Refactor, use outbound socket also for incoming requests
The Achilles Cyber Security tester use inet port monitors to supervise a
device under test (DUT).  This simple SYN monitoring trigger a bug in
the design of ssdp-responder v1.5 and earlier.

The bug comes from ssdpd using separate sockets for receiving and
sending.  Only receiving sockets are poll()'ed and emptied on each
incoming request.  However, the sending sockets are also open for
external connections causing the kernel syn backlog to fill up to
max and then causing loss of function.

This patch is an initial refactor of the internal design to use one
socket per ifaddr and join 239.255.255.255 on each.  This should allow
us to receive and send SSDP frames on all interfaces, just like before
but without any of the side effects of the previous design.  Initial
tests are positive.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>