First really working ssdp-scan, with caching and progress spinner
This is a really nasty little tool.  It sends SSDP M-SEARCH * messages
to 239.255.255.250 every 100 msec, really stressing remote hosts.  When
it gets a response it fetches the XML using uget(), then parses it for
<friendlyName>foo</friendlyName>, <presentationURL>URI</presentationURL>
and then caches the results in a linked list of remote hosts to prevent
printing duplicates.  If the found host does not exist it is printed to
stdout and a new M-SEARCH * commences.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>