ssdpd: Fix unintentional periodic busy loop
The retry logic in wait_message is broken. The only things that will
cause the retry loop to exit is either the timeout reaching a negative
value or the underlying poll(2) returning with errno EINTR. This means
that if the underlying poll(2) hits the timeout and returns zero (no
file descriptors ready) wait_message will busy loop for a whole second
until the timeout goes from 0 to -1.

The only error condition where a retry should be reasonable is if the
poll was interrupted by a signal, but when a signal has occured we might
actually want to break out of the main loop to shutdown.

This patch simplifies the logic of the loop, to only continue to poll if
there is non-zero time left until timeout.

Signed-off-by: Jacques de Laval <Jacques.De.Laval@westermo.com>