If /etc/os-release does not have VERSION_ID, use VERSION_CODENAME
On Debian, while in a new release cycle (sid/testing), the VERSION_ID is
not set because it has not yet been released.  Use VERSION_CODENAME as a
fallback for Debian and other systems that don't set VERSION_ID.

We still fall back to lsb-release, or POSIX, if DISTRIB_ID is not found.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>